GENIO: soc: mediatek: cmdq: Fix pointer arithmetic
Casting a pointer to a u32 for the purpose of pointer arithmetic will
fail on 64-bit architectures if the pointer points to a location above
the 4GB limit. This is caught by the compiler, which produces a warning:

drivers/soc/mediatek/mtk-cmdq-helper.c: In function ‘cmdq_pkt_poll_timeout’:
drivers/soc/mediatek/mtk-cmdq-helper.c:795:60: error: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
  795 |                         inst = (struct cmdq_instruction *)((u32)inst + CMDQ_INST_SIZE);
      |                                                            ^
drivers/soc/mediatek/mtk-cmdq-helper.c:795:32: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]
  795 |                         inst = (struct cmdq_instruction *)((u32)inst + CMDQ_INST_SIZE);
      |                                ^

Fix it by casting it to uintptr_t.

+++ drivers/soc/mediatek/mtk-cmdq-helper.c
@@ -792,7 +792,7 @@ int cmdq_pkt_poll_timeout(struct cmdq_pkt *pkt, u32 value, u8 subsys,
 	if (cnt_end_addr_mark) {
 		inst = (struct cmdq_instruction *)(pkt->va_base + cnt_end_addr_mark);
 		if (inst && inst->op == CMDQ_CODE_JUMP)
-			inst = (struct cmdq_instruction *)((u32)inst + CMDQ_INST_SIZE);
+			inst = (struct cmdq_instruction *)((uintptr_t)inst + CMDQ_INST_SIZE);
 		shift_pa = (pkt->cmd_buf_size - cnt_end_addr_mark - CMDQ_INST_SIZE) >> shift_bits;
 		if (inst) {
 			inst->arg_b = CMDQ_GET_ARG_B(shift_pa);
